requirements.txt
streamlit
python-dotenv
openai>=1.0.0
google-api-python-client
py-expression
